requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-debug-toolbar==3.2.4
django-dotenv==1.4.2
django-upgrade==1.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filter==0.0.0.20200724
httpie==3.0.2
idna==3.3
importlib-metadata==4.10.1
Markdown==3.3.6
multidict==6.0.2
oauthlib==3.2.0
psycopg2==2.9.3
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
tokenize-rt==4.2.1
urllib3==1.26.8
zipp==3.7.0
